Raise an exception, don't just raise
The "raise" syntax is only used in an exception handler
to re-raise the original exception.  Otherwise, "raise"
raises an exception that you're raising None:
TypeError: exceptions must be old-style classes or derived from BaseException, not NoneType
which does cause the program to exit but does not provide
any useful information about why.